Updating with all current changes
As it currently stands the best first search is performing worse than expected
Analysis shows that while it starts well, it quickly begins picking paths that aren't good
The solution is to rework the inside of the search algorithm and change its scoring.
Scoring is what is causing the main issue, since it is switching to a worse path and not exiting it.